+++ CHANGES.md
@@ -5,6 +5,8 @@
 <!-- PR authors:
      Please include the PR number in the changelog entry, not the issue number -->
 
+- Add support for NO_COLOR environment variable to disable ANSI output (#5129)
+
 ### Highlights
 
 <!-- Include any especially major or disruptive changes here -->